# Improve speed of Base58 Encoding