@laanwj 2021-02-15
The gains are quite impressive!

That said, this does make the code longer and quite a bit more complicated.

What motivated you to optimize this specifically? Are you experiencing any behavior where decoding / encoding Base58 is in the critical path?