@martinus 2021-02-15
In my tool [BitcoinUtxoVisualizer](https://github.com/martinus/BitcoinUtxoVisualizer) I am basically limited by the speed of how fast bitcoind can send me JSON block data. While profiling I saw `EncodeBase58` quite a few times so I had a look at it. It's about 2% faster or so with that change.

So for my use case honestly this change is probably not worth it. There are other much more important changes, e.g. #21006. But I saw @promag's PR #7656 so having a faster Base58 encoding might be worthwhile for others? If not, feel free to decline the PR, I'm perfectly fine with that!