@laanwj 2021-02-15
> In my tool BitcoinUtxoVisualizer I am basically limited by the speed of how fast bitcoind can send me JSON block data

Thanks, that makes sense.

> might be worthwhile for others? If not, feel free to decline the PR, I'm perfectly fine with that!

I'll leave that to others. If this can get enough review I'm fully okay with it! I'm not really declining this, my point was only that "reviewable straightforward code" is important from a security perspective and this sometimes trumps pure performance (depending on the part of the code). Especially in a memory-unsafe language like C++ where it's so easy to reach outside a buffer. We've had some discussions about this in the past when Base58 seemed to attract a lot of pretty arbitrary "make it faster" changes. At least you're doing proper benchmarking.