@practicalswift 2021-02-15
Personally I find the existing version in `master` much easier to reason about.

I tend to prefer "dumb and easy to reason about" over "cleverly optimised but harder to reason about" in code that is not performance critical.

I don't mean to be boring: this is a clever optimization. I'm just afraid it removes the "trivially correct" property of the code it optimizes :)