@martinus 2021-02-15
I agree the existing version is easier to read. I did not look at IBD at all, are there still improvements to be done there? It seemed to me that it's already well optimized.

I mostly profiled fetching JSON block data. The biggest optimization opportunity for this is reducing the global lock, and making univalue faster.