@martinus 2021-02-17
Closing, as it's not performance relevant but critical code so more complexity is bad. Thanks for having a look!